Fix: 서버 시간 테스트 페이지 리팩토링 및 기존 코드 삭제 #93
